+++ tests/aux_search/test_main.py
@@ -30,7 +30,7 @@ def test_wikidata_search_and_import_results(aux_page: Page) -> None:
     expect(search_input).to_be_visible()
 
     # test search input is showing correctly
-    search_input.fill("this doesn't exist gs871s9j91kksWN0shx345jnj")
+    search_input.fill("Q000000")  # does not exist
     search_input.press("Enter")
     page.screenshot(path="tests_aux_search_test_main-test_search-input-0-found.png")
     expect(page.get_by_text("showing 0 of")).to_be_visible()
@@ -38,9 +38,7 @@ def test_wikidata_search_and_import_results(aux_page: Page) -> None:
     # test expand button works
     search_input.fill("Q12345")
     search_input.press("Enter")
-    expand_all_properties_button = page.get_by_test_id("expand-properties-button").nth(
-        1
-    )
+    expand_all_properties_button = page.get_by_test_id("expand-properties-button").first
     page.screenshot(path="tests_aux_search_test_main-search_result.png")
     expect(page.get_by_text("Count von Count")).to_be_visible()
     expect(page.get_by_test_id("all-properties-display")).not_to_be_visible()
@@ -49,21 +47,28 @@ def test_wikidata_search_and_import_results(aux_page: Page) -> None:
     page.screenshot(path="tests_aux_search_test_main-test_expand_button.png")
 
     # test import button works
-    import_button = page.get_by_text("Import").nth(1)
+    import_button = page.get_by_text("Import").first
     import_button.click()
-    expect(page.get_by_text("Aux search result imported successfully")).to_be_visible()
+    toast = page.locator(".editor-toast").first
+    expect(toast).to_be_visible()
+    expect(toast).to_contain_text("Organization was imported successfully.")
     expect(import_button).to_be_disabled()
     page.screenshot(path="tests_aux_search_test_main-test_import_button.png")
 
     # test node was ingested into backend
     connector = BackendApiConnector.get()
-    result = connector.fetch_extracted_items(q="Count von Count")
-    assert result["total"] >= 1
+    result = connector.fetch_extracted_items("Count von Count", None, None, 0, 1)
+    assert result.total >= 1
 
 
 @pytest.mark.integration
 @pytest.mark.external
 def test_ldap_search_and_import_results(aux_page: Page) -> None:
+    # count the persons before
+    connector = BackendApiConnector.get()
+    result = connector.fetch_extracted_items(None, None, ["ExtractedPerson"], 0, 1)
+    persons_before = result.total
+
     page = aux_page
     ldap_tab = page.get_by_role("tab", name="LDAP")
     ldap_tab.click()
@@ -79,11 +84,9 @@ def test_ldap_search_and_import_results(aux_page: Page) -> None:
     )
 
     # test expand button works
-    search_input.fill("Ciftci*")
+    search_input.fill("L*")
     search_input.press("Enter")
-    expand_all_properties_button = page.get_by_test_id("expand-properties-button").nth(
-        1
-    )
+    expand_all_properties_button = page.get_by_test_id("expand-properties-button").first
     page.screenshot(path="tests_aux_search_test_main-search_result_lap.png")
     expect(page.get_by_text("Ciftci, Zehra")).to_be_visible()
     expect(page.get_by_test_id("all-properties-display")).not_to_be_visible()
@@ -92,16 +95,17 @@ def test_ldap_search_and_import_results(aux_page: Page) -> None:
     page.screenshot(path="tests_aux_search_test_main-test_ldap_expand_button.png")
 
     # test import button works
-    import_button = page.get_by_text("Import").nth(1)
+    import_button = page.get_by_text("Import").first
     import_button.click()
-    expect(page.get_by_text("Aux search result imported successfully")).to_be_visible()
+    toast = page.locator(".editor-toast").first
+    expect(toast).to_be_visible()
+    expect(toast).to_contain_text("Person was imported successfully.")
     expect(import_button).to_be_disabled()
     page.screenshot(path="tests_aux_search_test_main-test_ldap_import_button.png")
 
-    # test node was ingested into backend
-    connector = BackendApiConnector.get()
-    result = connector.fetch_extracted_items(q="Ciftci")
-    assert result["total"] >= 1
+    # count the persons after
+    result = connector.fetch_extracted_items(None, None, ["ExtractedPerson"], 0, 1)
+    assert result.total == persons_before + 1
 
 
 @pytest.mark.integration
@@ -141,8 +145,8 @@ def test_orcid_search_and_import_results(aux_page: Page) -> None:
 
     # test node was ingested into backend
     connector = BackendApiConnector.get()
-    result = connector.fetch_extracted_items(q="Kamran")
-    assert result["total"] >= 1
+    result = connector.fetch_extracted_items("Kamran", None, None, 0, 1)
+    assert result.total >= 1
 
 
 @pytest.mark.integration