+++ tests/test_transform.py
@@ -1,6 +1,5 @@
 import pytest
 
-from mex.common.exceptions import MExError
 from mex.common.models import AnyExtractedModel
 from mex.common.types import APIType, Identifier, Link, LinkLanguage, Text, TextLanguage
 from mex.editor.models import EditorValue
@@ -67,7 +66,9 @@ def test_transform_values(
 
 
 def test_transform_value_none_error() -> None:
-    with pytest.raises(MExError, match="cannot transform NoneType to editor value"):
+    with pytest.raises(
+        NotImplementedError, match="cannot transform NoneType to editor value"
+    ):
         transform_value(None)
 
 
@@ -110,6 +111,10 @@ def test_transform_models_to_title(dummy_data: list[AnyExtractedModel]) -> None:
             # activity renders title as text (with language badge)
             EditorValue(text="Aktivität 1", badge="DE")
         ],
+        [
+            # resource renders title as text
+            EditorValue(text="Bioinformatics Resource 1"),
+        ],
     ]
 
 
@@ -127,25 +132,16 @@ def test_transform_models_to_preview(dummy_data: list[AnyExtractedModel]) -> Non
         [EditorValue(text="Unit 1", badge="EN", enabled=True)],
         [
             EditorValue(text="A1", enabled=True),
-            EditorValue(
-                identifier="wEvxYRPlmGVQCbZx9GAbn",
-            ),
-            EditorValue(
-                identifier="g32qzYNVH1Ez7JTEk3fvLF",
-            ),
-            EditorValue(
-                identifier="cWWm02l1c6cucKjIhkFqY4",
-            ),
-            EditorValue(
-                identifier="cWWm02l1c6cucKjIhkFqY4",
-            ),
-            EditorValue(
-                text="1999-12-24",
-                badge="day",
-            ),
-            EditorValue(
-                text="2023-01-01",
-                badge="day",
-            ),
+            EditorValue(identifier="wEvxYRPlmGVQCbZx9GAbn"),
+            EditorValue(identifier="g32qzYNVH1Ez7JTEk3fvLF"),
+            EditorValue(identifier="cWWm02l1c6cucKjIhkFqY4"),
+            EditorValue(identifier="cWWm02l1c6cucKjIhkFqY4"),
+            EditorValue(text="1999-12-24", badge="day"),
+            EditorValue(text="2023-01-01", badge="day"),
+        ],
+        [
+            EditorValue(identifier="cWWm02l1c6cucKjIhkFqY4"),
+            EditorValue(text="Theme", badge="BIOINFORMATICS_AND_SYSTEMS_BIOLOGY"),
+            EditorValue(text="AccessRestriction", badge="OPEN"),
         ],
     ]